+++ includes/gookto_io/intersection/intersectionPhase.hpp
@@ -8,20 +8,20 @@ class intersectionPhase
 private:
     int id;
     std::vector<int> connectionRef; 
-    std::vector<float> priority;
+    std::vector<double> priority;
 
 
 public:
     intersectionPhase(int id);
 
     void pushConnectionRef(int val);
-    void pushPriority(float val);
+    void pushPriority(double val);
 
     //return connectionRef and Priority
 
     int getId() { return id; }
     std::vector<int> getConnRef() { return connectionRef; }
-    std::vector<float> getPriority() { return priority; }
+    std::vector<double> getPriority() { return priority; }
 
 
 };
